Theme: moved defaultTheme to separate constant file
