+++ client/packages/lowcoder/src/comps/controls/styleControlConstants.tsx
@@ -2,6 +2,7 @@ import {ThemeDetail} from "api/commonSettingApi";
 import {darkenColor, isDarkColor, lightenColor, toHex} from "lowcoder-design";
 import {trans} from "i18n";
 import {StyleConfigType} from "./styleControl";
+import { defaultTheme } from "@lowcoder-ee/constants/themeConstants";
 
 type SupportPlatform = "pc" | "mobile";
 
@@ -224,30 +225,6 @@ export type SingleColorConfig =
   | BoxShadowConfig
   | BoxShadowColorConfig
   | AnimationIterationCountConfig;
-
-export const defaultTheme: ThemeDetail = {
-  primary: "#3377FF",
-  textDark: "#222222",
-  textLight: "#FFFFFF",
-  canvas: "#F5F5F6",
-  primarySurface: "#FFFFFF",
-  borderColor: "#D7D9E0",
-  borderRadius: "4px",
-  borderWidth: "1px",
-  borderStyle: "solid",
-  margin: "3px",
-  padding: "3px",
-  gridColumns: "24",
-  textSize: "14px",
-  animation: "",
-  animationDelay: "",
-  animationDuration: "",
-  opacity: "1",
-  boxShadow: "",
-  boxShadowColor: "",
-  animationIterationCount: "",
-  components: {},
-};
 
 export const SURFACE_COLOR = "#FFFFFF";
 const SECOND_SURFACE_COLOR = "#D7D9E0";
